move node stuff into engine/node.cpp/hpp
