+++ CMakeLists.txt
@@ -28,6 +28,8 @@ set(OpenGL_GL_PREFERENCE GLVND)
 add_library(engine STATIC
     engine/raybuff.cpp
     engine/raybuff.hpp
+    engine/node.cpp
+    engine/node.hpp
     engine/scene.cpp
     engine/scene.hpp
     engine/core.cpp
